requirements.txt
streamlit
streamlit-extras
streamlit_option_menu
opencv-python>=4.9.0.80
numpy>=1.26.4
Pillow>=10.2.0
snowflake-connector-python>=3.7.0
ultralytics>=8.1.0
huggingface-hub>=0.20.0
openai>=1.12.0
pandas>=2.2.0
plotly>=5.18.0
setuptools<71
langchain_community
langchain_openai
langchain_core
langchain==0.1.16
tiktoken
pytz